[FEAT] Allows contexts to be sparse arrays (#62)
* FIX: make `groupby_array` sparse-compatible

Use np.split on a single index array rather than using it on each array to be grouped. This means groups are sliced with fancy indexing (which works with sparse matrices) rather than np.split, which does not.

* FEAT: Allows context to be sparse for Normal regressors

Mostly just assumes the user knows what they're doing if they've provided a sparse context. Providing a sparse context skips several validation checks, under the belief that you'd only do that if you have a sparse learner.
Fixes #61

* TEST: expand tests to ensure that sparse contexts result in the same learned model as dense contexts.

* MAINT: use `csc_array` instead of `csc_matrix`

This is recommended by scipy.sparse for future compatibility changes (and sparse arrays are easier to swap with numpy arrays)

* DOCS: re-run doc notebooks

* DOCS: remove mentions of sparse matrices

* DOCS: sparse array