Enhance support for custom policies (#117)
* Adds a generic type parameter P to `Agent` and `ContextualAgent` that's bounded by `PolicyProtocol`

This enhances the ability to write custom policy objects and have their concrete type detectable by type checkers (as opposed to just the policy interface). This allows, for example, a static type checker to detect a custom policy that has a `mask_arms` method that can be accessed from `agent.policy.mask_arms(...)`. As such, this fixes #102

* TEST: add types for tests